+++ source/atofigh_adapter/atofigh.cpp
@@ -14,11 +14,11 @@ using namespace arbok;
 
 struct arbok::AtofighImpl {
     AtofighImpl(int n, int m) { }
-    void create_edge(int from, int to, int weight) {
+    void create_edge(int from, int to, double weight) {
         m_edges.emplace_back(from,to);
         m_weights.push_back(weight);
     }
-    long long run(int root) {
+    long double run(int root) {
         array<int,1> roots{root};
         auto boost_edge_list = boost::edge_list(begin(m_edges),end(m_edges));
         edmonds_optimum_branching<false,true,true>(
@@ -28,7 +28,7 @@ struct arbok::AtofighImpl {
                 begin(roots),
                 end(roots),
                 std::back_inserter(m_arbo));
-        long long ans = 0;
+        long double ans = 0;
         for(auto e_id : m_arbo)
             ans += m_weights[e_id];
         return ans;
@@ -38,15 +38,15 @@ struct arbok::AtofighImpl {
     }
 
     std::vector<std::pair<int,int>> m_edges;
-    std::vector<long long> m_weights;
+    std::vector<long double> m_weights;
     std::vector<int> m_arbo;
 };
 
 // the adapter just passes through to the impl
 Atofigh::~Atofigh() = default;
 arbok::Atofigh::Atofigh(int n, int m) : num_vertices(n), m_impl(make_unique<AtofighImpl>(n, m)) { }
-void Atofigh::create_edge(int from, int to, int weight) { m_impl->create_edge(from,to,weight); }
-long long Atofigh::run(int root) { return m_impl->run(root); }
+void Atofigh::create_edge(int from, int to, double weight) { m_impl->create_edge(from,to,weight); }
+long double Atofigh::run(int root) { return m_impl->run(root); }
 std::vector<int> Atofigh::reconstruct(int root) { return m_impl->reconstruct(root); }
 
 #else
@@ -58,7 +58,7 @@ arbok::Atofigh::Atofigh(int n, int m) : num_vertices(n) {
     std::cout << "ERROR: can't use lemon solver without lemon" << endl;
     exit(1);
 }
-void Atofigh::create_edge(int from, int to, int weight) { }
-long long Atofigh::run(int root) { return 0; }
+void Atofigh::create_edge(int from, int to, double weight) { }
+long double Atofigh::run(int root) { return 0; }
 std::vector<int> Atofigh::reconstruct(int root) { return {}; }
 #endif
